+++ citywok_ms/models.py
@@ -78,3 +78,21 @@ class Employee(db.Model):
             raise ValueError(f'"{sex}" not in SEX set')
         else:
             return sex
+
+
+class Supplier(db.Model):
+    id = Column(Integer, primary_key=True)
+    name = Column(String, nullable=False)
+    abbreviation = Column(String)
+    principal = Column(String)
+    contact = Column(Integer)
+    email = Column(String)
+    nif = Column(Integer, unique=True)
+    iban = Column(String, unique=True)
+    address = Column(String)
+    postcode = Column(String)
+    city = Column(String)
+    remark = Column(Text)
+
+    def __repr__(self):
+        return f'Supplier({self.id}: {self.name})'